Modifying migration

+++ db/migrate/20210531023033_change_heigth_from_string_to_float_in_pet_history.rb
@@ -2,7 +2,7 @@ class ChangeHeigthFromStringToFloatInPetHistory < ActiveRecord::Migration[5.2]
   def change
     reversible do |dir|
       change_table :pet_histories do |t|
-        dir.up { t.change :heigth, :double }
+        dir.up { t.change :heigth, :float }
         dir.down { t.change :heigth, :string }
       end
     end
